Move demo to playground/framework for now, to test with and without ssr
