+++ playground/framework/app/routes.ts
@@ -2,5 +2,5 @@ import { type RouteConfig, index, route } from "@react-router/dev/routes";
 
 export default [
   index("routes/_index.tsx"),
-  route("products/:id", "routes/product.tsx"),
+  route("images/:id", "routes/image.tsx"),
 ] satisfies RouteConfig;
